test(composer-app): refactor playwright tests to run more concurrently
