+++ packages/apps/composer-app/src/playwright/plugins/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2024 DXOS.org
+//
+
+export * from './markdown';
+export * from './stack';
